Swap a duplicate
